src/functions/AnalyzeInvoiceProcess/requirements.txt
tqdm==4.66.1
azure-functions==1.17.0
